requirements.txt
streamlit
langchain
langchain-groq
python-dotenv
pypdf
langchain-google-genai
chromadb
